Crate nannou_osc

Tools for working with OSC. sender() creates an OSC sender, receiver(port) creates an OSC receiver.

Re-exports

pub use self::recv::Receiver;

pub use self::send::Sender;

pub use rosc;

Modules

address

Address checking and matching methods

decoder

Provides a decoding method for OSC packets.

encoder

Encodes an OscPacket to a byte vector.

recv

Items related to the osc::Receiver implementation.

send

Items related to the osc::Sender implementation.

Structs

Array

An OscArray color.

Bundle

An OSC bundle contains zero or more OSC packets and a time tag. The contained packets should be applied at the given time tag.

Color

An RGBA color.

Connected

Indicates that a Sender is currently connected to a specific target address.

Message

An OSC message consists of an address and zero or more arguments. The address should specify an element of your Instrument (or whatever you want to control with OSC) and the arguments are used to set properties of the element to the respective values.

MidiMessage

Represents the parts of a Midi message. Mainly used for tunneling midi over a network using the OSC protocol.

Time

A time tag in OSC message consists of two 32-bit integers where the first one denotes the number of seconds since 1900-01-01 and the second the fractions of a second. For details on its semantics see http://opensoundcontrol.org/node/3/#timetags

TimeError

An error returned by conversions involving OscTime.

Unconnected

Indicates that a Sender is not currently connected to a target address, and that the target address will have to be supplied manually when sending packets.

Enums

CommunicationError

Errors that might occur whilst attempting to send or receive an OSC packet.

Error

Represents errors returned by decode or encode.

Packet

An OSC packet can contain an OSC message or a bundle of nested packets which is called an OSC bundle.

Type

see OSC Type Tag String: OSC Spec. 1.0 padding: zero bytes (n*4)

Functions

decode

Decodes the given slice of bytes into a Packet.

default_ipv4_addr

The default local IP address.

encode

Encodes the given Packet into a Vec of bytes.

msg

A simplified constructor for an OSC Message.

receiver

A simple wrapper around the most commonly used Receiver constructor.

sender

A simple wrapper around the most commonly used Sender constructor.